Copia del backend a mi rama

+++ src/main/java/com/example/tuFondaOnline/service/PublicacionService.java
@@ -3,7 +3,9 @@ package com.example.tuFondaOnline.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.example.tuFondaOnline.model.Publicacion;
 import com.example.tuFondaOnline.repository.PublicacionRepository;
@@ -22,11 +24,13 @@ public class PublicacionService {
     }
     //Buscar una publicación con id
     public Publicacion findById(Long id) {
-        return publicacionRepository.findById(id).orElse(null);
+        return publicacionRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Publicación no encontrada"));
     }
     //Eliminar publicación
     public void delete(Long id) {
+        if (!publicacionRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se puede eliminar, ID no existe");
+        }
         publicacionRepository.deleteById(id);
     }
 }
-
